Updated speedtest to allow it to run as non-root user when run by cmk
I didn't want to run the 'speedtest-cli' Python script as root, so I  modified it to run as a user named 'cmkplugin' (which has to be  manually created on the host). The user doesn't need any special privileges.

If the script is run as a user other than root, it will continue to run as that user (for testing from the command line)

Changes to be committed:
-	modified:   speedtest